+++ public/manifest.json
@@ -9,19 +9,19 @@
   "start_url": "/",
   "icons": [
     {
-      "src": "/images/sobok-logo-144.webp",
+      "src": "/images/icon-144.webp",
       "sizes": "144x144",
       "type": "image/webp",
       "purpose": "maskable"
     },
     {
-      "src": "/images/sobok-logo-192.webp",
+      "src": "/images/icon-192.webp",
       "sizes": "192x192",
       "type": "image/webp",
       "purpose": "maskable"
     },
     {
-      "src": "/images/sobok-logo-512.webp",
+      "src": "/images/icon-512.webp",
       "sizes": "512x512",
       "type": "image/webp"
     }
