+++ src/layouts/SearchFormLayout.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image'
 import { ReactNode } from 'react'
 import ClientSideLink from 'src/components/atoms/ClientSideLink'
+import { TABLET_MIN_WIDTH, TOP_HEADER_HEIGHT } from 'src/models/constants'
 import BackArrowIcon from 'src/svgs/BackArrowIcon'
-import { TABLET_MIN_WIDTH, TOP_HEADER_HEIGHT } from 'src/utils/constants'
 import styled from 'styled-components'
 
 const HomeContainer = styled.div`
